Replaced the Login and Register buttons on the Navbar to a single Sign In Button. When clicking on the Sign In button, it pops up a dialog window that includes:
-- a close window button at the top right
-- a username, email, and password text entry box
-- a submit button, that will send the user information to App.js

I thought that might be better instead of a dedicated login/register
page since a user won't have to leave the page they are currently on.
All they have to do is either press the close button or submit log in
details to either go back to or update their current page.